Fix little-endian byte-order input encryption and decryption.
This fixes bug <https://github.com/jashandeep-sohi/python-blowfish/issues/9>.

`u4_1_struct` is used to split 32-bit integers into 4 bytes. The higher 8 bits
should be the first byte, the next 8 bits should be the next byte, and so on.
So, it should always be in big-endian byte order and not be dependent on the
`byte_order` option.